refactor: move team management views to module
